Minor restructuring, made source files and templates be packaged
