+++ Sources/Props/SourceGenerator.Build.props
@@ -14,6 +14,21 @@
     <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
   </ItemGroup>
 
+  <!-- Embed all resource (template and cs template) files -->
+  <ItemGroup>
+    <Content Include="$(MSBuildProjectDirectory)\Templates\**\*.*">
+      <Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
+      <CopyToOutputDirectory>true</CopyToOutputDirectory>
+      <IncludeInPackage>true</IncludeInPackage>
+    </Content>
+
+    <Content Include="$(MSBuildProjectDirectory)\Sources\**\*.*">
+      <Link>%(RecursiveDir)%(Filename)%(Extension)</Link>
+      <CopyToOutputDirectory>true</CopyToOutputDirectory>
+      <IncludeInPackage>true</IncludeInPackage>
+    </Content>
+  </ItemGroup>
+
   <ItemGroup>
     <!-- Compile-time and analysis dependencies -->
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.9.0" PrivateAssets="All" />
@@ -21,11 +36,11 @@
     <PackageReference Include="Nullable" Version="1.3.0" PrivateAssets="All"/>
 
     <!-- Source-generator abstractions -->
-    <Compile Include="$(SolutionDir)/Shared/Libraries/SourceGenerator.Common/**/*.cs" Link="SourceGenerator.Common" />
+    <Compile Include="$(SolutionDir)/Shared/Libraries/SourceGenerator.Common/**/*.cs" Link="SourceGenerator.Common" Visible="False" />
     <Compile Remove="$(SolutionDir)/Shared/Libraries/SourceGenerator.Common/obj/**/*.cs" />
 
     <!-- Run-time polyfill libraries -->
-    <Compile Include="$(SolutionDir)/Shared/Libraries/Polyfill/**/*.cs" Link="Polyfill" />
+    <Compile Include="$(SolutionDir)/Shared/Libraries/Polyfill/**/*.cs" Link="Polyfill" Visible="False" />
     <Compile Remove="$(SolutionDir)/Shared/Libraries/Polyfill/obj/**/*.cs" />
 
     <PackageReference Include="System.Memory" Version="4.5.4" PrivateAssets="all" GeneratePathProperty="true" />
