+++ Sources/SynKit/Libraries/Lexer.Generator/Lexer.Generator.csproj
@@ -2,10 +2,15 @@
   <Import Project="..\..\..\Props\SourceGenerator.Build.props" />
 
   <ItemGroup>
-    <Compile Include="$(SolutionDir)/Shared\Libraries/Collections/**/*.cs" Link="Collections" />
+    <Compile Include="$(SolutionDir)/Shared\Libraries/Collections/**/*.cs" Link="Collections" Visible="False" />
     <Compile Remove="$(SolutionDir)/Shared\Libraries/Collections/obj/**/*.cs" />
 
-    <Compile Include="$(SolutionDir)/SynKit\Libraries/Automata/**/*.cs" Link="Automata" />
+    <Compile Include="$(SolutionDir)/SynKit\Libraries/Automata/**/*.cs" Link="Automata" Visible="False" />
     <Compile Remove="$(SolutionDir)/SynKit\Libraries/Automata/obj/**/*.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Sources\" />
+    <Folder Include="Templates\" />
   </ItemGroup>
 </Project>
