+++ Sources/SynKit/Libraries/Lexer.Generator/Templates/lexer-template.sbncs
